requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
cloudinary==1.44.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
dj-database-url==2.3.0
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.33.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
PyYAML==6.0.2
redis==6.1.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
service-identity==24.2.0
setuptools==80.7.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.3
tokenizers==0.21.2
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
transformers==4.53.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
zope.interface==7.2
